+++ Yosemite/Yosemite/Model/Storage/OrderItemTaxRefund+ReadOnlyConvertible.swift
@@ -9,7 +9,7 @@ extension Storage.OrderItemTaxRefund: ReadOnlyConvertible {
     /// Updates the Storage.OrderItemTax with the ReadOnly.
     ///
     public func update(with orderItemTaxRefund: Yosemite.OrderItemTaxRefund) {
-        taxID = Int64(orderItemTaxRefund.taxID)
+        taxID = orderItemTaxRefund.taxID
         subtotal = orderItemTaxRefund.subtotal
         total = orderItemTaxRefund.total
     }
@@ -17,7 +17,7 @@ extension Storage.OrderItemTaxRefund: ReadOnlyConvertible {
     /// Returns a ReadOnly version of the receiver.
     ///
     public func toReadOnly() -> Yosemite.OrderItemTaxRefund {
-        return OrderItemTaxRefund(taxID: Int64(taxID),
+        return OrderItemTaxRefund(taxID: taxID,
                                   subtotal: subtotal ?? "",
                                   total: total ?? "")
     }
