+++ Yosemite/Yosemite/Model/Storage/OrderRefundCondensed+ReadOnlyConvertible.swift
@@ -9,7 +9,7 @@ extension Storage.OrderRefundCondensed: ReadOnlyConvertible {
     /// Updates the Storage.OrderRefundCondensed with the ReadOnly.
     ///
     public func update(with orderRefundCondensed: Yosemite.OrderRefundCondensed) {
-        refundID = Int64(orderRefundCondensed.refundID)
+        refundID = orderRefundCondensed.refundID
         reason = orderRefundCondensed.reason
         total = orderRefundCondensed.total
     }
@@ -17,7 +17,7 @@ extension Storage.OrderRefundCondensed: ReadOnlyConvertible {
     /// Returns a ReadOnly version of the receiver.
     ///
     public func toReadOnly() -> Yosemite.OrderRefundCondensed {
-        return OrderRefundCondensed(refundID: Int64(refundID),
+        return OrderRefundCondensed(refundID: refundID,
                                     reason: reason ?? "",
                                     total: total ?? "")
     }
