+++ Yosemite/Yosemite/Model/Storage/ShipmentTracking+ReadOnlyConvertible.swift
@@ -9,8 +9,8 @@ extension Storage.ShipmentTracking: ReadOnlyConvertible {
     /// Updates the Storage.SiteSetting with the a ReadOnly.
     ///
     public func update(with shipmentTracking: Yosemite.ShipmentTracking) {
-        siteID = Int64(shipmentTracking.siteID)
-        orderID = Int64(shipmentTracking.orderID)
+        siteID = shipmentTracking.siteID
+        orderID = shipmentTracking.orderID
         trackingID = shipmentTracking.trackingID
         trackingNumber = shipmentTracking.trackingNumber
         trackingProvider = shipmentTracking.trackingProvider
@@ -21,8 +21,8 @@ extension Storage.ShipmentTracking: ReadOnlyConvertible {
     /// Returns a ReadOnly version of the receiver.
     ///
     public func toReadOnly() -> Yosemite.ShipmentTracking {
-        return ShipmentTracking(siteID: Int64(siteID),
-                                orderID: Int64(orderID),
+        return ShipmentTracking(siteID: siteID,
+                                orderID: orderID,
                                 trackingID: trackingID,
                                 trackingNumber: trackingNumber ?? "",
                                 trackingProvider: trackingProvider,
